+++ src/Stormancer.Plugins/GameFinder/Stormancer.Server.Plugins.GameFinder/PartyGameFinder/PartyGameFinderResolver.cs
@@ -91,7 +91,7 @@ namespace Stormancer.Server.Plugins.GameFinder
                     var party = gameCtx.Game.AllParties.First();
                     var leaderId = party.PartyLeaderId;
                     var leader = party.Players[leaderId];
-                    config.HostSessionId = leader.SessionId;
+                    config.HostSessionId = leader.SessionId.ToString();
                 }